Move the JVM anchor into the wiring model.
