+++ platform-sdk/swirlds-common/src/main/java/com/swirlds/common/wiring/model/internal/StandardWiringModel.java
@@ -22,6 +22,9 @@ import static com.swirlds.common.wiring.schedulers.builders.TaskSchedulerType.DI
 import static com.swirlds.common.wiring.schedulers.builders.TaskSchedulerType.SEQUENTIAL_THREAD;
 
 import com.swirlds.common.context.PlatformContext;
+import com.swirlds.common.threading.locks.AutoClosableLock;
+import com.swirlds.common.threading.locks.internal.AutoLock;
+import com.swirlds.common.threading.locks.locked.Locked;
 import com.swirlds.common.wiring.model.ModelEdgeSubstitution;
 import com.swirlds.common.wiring.model.ModelGroup;
 import com.swirlds.common.wiring.model.ModelManualLink;
@@ -103,6 +106,13 @@ public class StandardWiringModel implements WiringModel {
     private boolean started = false;
 
     /**
+     * Used to protect access to the JVM anchor.
+     */
+    private final AutoClosableLock anchorLock = new AutoLock();
+
+    private JvmAnchor anchor;
+
+    /**
      * Constructor.
      *
      * @param platformContext the platform context
@@ -143,6 +153,32 @@ public class StandardWiringModel implements WiringModel {
     public OutputWire<Instant> buildHeartbeatWire(final double frequency) {
         throwIfStarted();
         return getHeartbeatScheduler().buildHeartbeatWire(frequency);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void lowerJvmAnchor() {
+        try (final Locked ignored = anchorLock.lock()) {
+            if (anchor == null) {
+                anchor = new JvmAnchor();
+                anchor.start();
+            }
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public void raiseJvmAnchor() {
+        try (final Locked ignored = anchorLock.lock()) {
+            if (anchor != null) {
+                anchor.stop();
+                anchor = null;
+            }
+        }
     }
 
     /**
@@ -371,6 +407,8 @@ public class StandardWiringModel implements WiringModel {
         for (final SequentialThreadTaskScheduler<?> threadScheduler : threadSchedulers) {
             threadScheduler.stop();
         }
+
+        raiseJvmAnchor();
     }
 
     /**
