+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/SwirldsPlatform.java
@@ -70,7 +70,6 @@ import com.swirlds.platform.components.SavedStateController;
 import com.swirlds.platform.components.appcomm.DefaultLatestCompleteStateNotifier;
 import com.swirlds.platform.components.appcomm.LatestCompleteStateNotifier;
 import com.swirlds.platform.config.StateConfig;
-import com.swirlds.platform.config.ThreadConfig;
 import com.swirlds.platform.config.TransactionConfig;
 import com.swirlds.platform.consensus.ConsensusSnapshot;
 import com.swirlds.platform.consensus.NonAncientEventWindow;
@@ -784,10 +783,6 @@ public class SwirldsPlatform implements Platform {
                         Pair.of(platformWiring, "platformWiring"),
                         Pair.of(shadowGraph, "shadowGraph"),
                         Pair.of(transactionPool, "transactionPool")));
-
-        if (platformContext.getConfiguration().getConfigData(ThreadConfig.class).jvmAnchor()) {
-            thingsToStart.add(new JvmAnchor(threadManager));
-        }
     }
 
     /**
@@ -1010,6 +1005,7 @@ public class SwirldsPlatform implements Platform {
     @Override
     public void start() {
         logger.info(STARTUP.getMarker(), "Starting platform {}", selfId);
+        platformWiring.getModel().lowerJvmAnchor();
 
         thingsToStart.start();
